Updated the visual simulation
Devices now have velocities so that we can update them as we wish while each tick still moves the devices automatically. Also, made the simulation act the same regardless of FPS. I tested with 4, 8, 30, and 60 FPS, and the simulation cannot handle 30 or 60 FPS on my computer (probably an issue with the device tick being expensive).